pyproject.toml
[project]
name = "goit-pythonweb-hw-12"
version = "0.1.0"
description = ""
authors = [
    {name = "maksar24",email = "74876375+maksar24@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.16.1,<2.0.0)",
    "pydantic[email] (>=2.11.5,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "libgravatar (>=1.0.4,<2.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "fastapi-mail (>=1.5.0,<2.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "cloudinary (>=1.44.1,<2.0.0)",
    "bcrypt (<4.1.0)",
    "redis[asyncio] (>=6.2.0,<7.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "pytest-cov (>=6.2.1,<7.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "asgi-lifespan (>=2.1.0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "src"}]


[tool.poetry.group.dev.dependencies]
sphinx = "^8.2.3"
pytest-asyncio = "^1.0.0"

[tool.pytest.ini_options]
markers = [
    "asyncio: mark async tests"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
